pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "my_ai_services_sdk"
version = "0.1.1" # 習慣上，修復 bug 或小調整時，可以增加最後一位版本號
authors = [
  { name="Your Name", email="your.email@example.com" },
]
description = "A personal SDK for accessing AI services like Azure and Gemini."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "azure-cognitiveservices-speech",
    "google-generativeai"
]

[tool.setuptools.packages.find]
where = ["src"]  # <--- 把 "." 改成 "src"
include = ["my_ai_services_sdk*"]
exclude = ["tests*"]